refactor: schema 따로 분리
