+++ packages/api/src/inbox/index.ts
@@ -1,1 +1,2 @@
 export * from './getAllUnreadInbox';
+export * from './schema';
